bsp: u-boot: deploy separated DT and u-boot binary
U-boot is now configured with CONFIG_OF_SEPARATE.

Unfortunately, we cannot confibure separted vs merged boot using env
alone (because the DT address, that we want to control due to alignment,
can only be hardcoded in the *early* env). So, if we want to support
separated boot, then we have to also make it the default.